[Refactor] #36 - AddScheduleView UI 구조 개선
PickerButton 분리
변수명 변경
